requirements.txt
opencv-python
matplotlib
pandas
numpy
Pillow
mtcnn
keras
scikit-learn

